Move get cohort_definition methods to CohortDefinitionService
