+++ src/main/java/org/ohdsi/webapi/service/CohortDefinitionService.java
@@ -5,20 +5,31 @@
  */
 package org.ohdsi.webapi.service;
 
+import java.sql.ResultSet;
+import java.sql.SQLException;
+import java.util.List;
+
 import javax.servlet.ServletContext;
 import javax.ws.rs.Consumes;
+import javax.ws.rs.GET;
 import javax.ws.rs.POST;
 import javax.ws.rs.Path;
+import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
 import javax.ws.rs.core.Context;
 import javax.ws.rs.core.MediaType;
+
 import org.ohdsi.sql.SqlRender;
 import org.ohdsi.sql.SqlTranslate;
-
 import org.ohdsi.webapi.cohortdefinition.CohortExpression;
 import org.ohdsi.webapi.cohortdefinition.CohortExpressionQueryBuilder;
+import org.ohdsi.webapi.helper.ResourceHelper;
+import org.ohdsi.webapi.model.CohortDefinition;
 import org.ohdsi.webapi.sqlrender.TranslatedStatement;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Component;
 
 /**
@@ -34,11 +45,43 @@ public class CohortDefinitionService {
   @Context
   ServletContext context;
 
-    @Value("${datasource.cdm.schema}")
-    private String cdmSchema;
-    
-    @Value("${datasource.dialect}")
-    private String dialect;
+  @Value("${datasource.cdm.schema}")
+  private String cdmSchema;
+  
+  @Value("${datasource.ohdsi.schema}")
+  private String ohdsiSchema;
+  
+  @Value("${datasource.dialect}")
+  private String dialect;
+  
+  @Value("${datasource.dialect.source}")
+  private String sourceDialect;
+  
+  private final JdbcTemplate jdbcTemplate;
+  
+  /**
+   * Constructor needed for Spring
+   * 
+   * @param jdbcTemplate
+   */
+  @Autowired
+  public CohortDefinitionService(final JdbcTemplate jdbcTemplate) {
+      this.jdbcTemplate = jdbcTemplate;
+  }
+  
+  private final RowMapper<CohortDefinition> cohortDefinitionMapper = new RowMapper<CohortDefinition>() {
+      
+      @Override
+      public CohortDefinition mapRow(final ResultSet rs, final int arg1) throws SQLException {
+          final CohortDefinition definition = new CohortDefinition();
+          definition.setCohortDefinitionDescription(rs.getString(CohortDefinition.COHORT_DEFINITION_DESCRIPTION));
+          definition.setCohortDefinitionId(rs.getInt(CohortDefinition.COHORT_DEFINITION_ID));
+          definition.setCohortDefinitionName(rs.getString(CohortDefinition.COHORT_DEFINITION_NAME));
+          definition.setCohortDefinitionSyntax(rs.getString(CohortDefinition.COHORT_DEFINITION_SYNTAX));
+          definition.setCohortInitiationDate(rs.getDate(CohortDefinition.COHORT_INITIATION_DATE));
+          return definition;
+      }
+  };
     
   @Path("generate")
   @POST
@@ -55,4 +98,42 @@ public class CohortDefinitionService {
     ts.targetSQL = translatedSql;
     return ts;
   }
+  
+  /**
+   * Returns all cohort definitions in the CDM schema, to be replaced 
+   * once there is a cohort_definition service
+   * 
+   * @return
+   */
+  @GET
+  @Path("/")
+  @Produces(MediaType.APPLICATION_JSON)
+  public List<CohortDefinition> getCohortDefinitionList() {
+      
+      String sql = ResourceHelper.GetResourceAsString("/resources/cohortdefinition/sql/getCohortDefinitions.sql");
+      sql = SqlRender.renderSql(sql, new String[] { "CDM_schema" }, new String[] { this.cdmSchema });
+      sql = SqlTranslate.translateSql(sql, this.sourceDialect, this.dialect);
+      
+      return this.jdbcTemplate.query(sql, this.cohortDefinitionMapper);
+  }
+  
+  /**
+   * @param id
+   * @return
+   */
+  @GET
+  @Path("/{id}")
+  @Produces(MediaType.APPLICATION_JSON)
+  public CohortDefinition getCohortDefinition(@PathParam("id") final int id) {
+      
+      String sql_statement = ResourceHelper.GetResourceAsString("/resources/cohortdefinition/sql/getCohortDefinitionsById.sql");
+      sql_statement = SqlRender.renderSql(sql_statement, new String[] { "id", "CDM_schema" },
+          new String[] { String.valueOf(id), this.cdmSchema });
+      sql_statement = SqlTranslate.translateSql(sql_statement, this.sourceDialect, this.dialect);
+      
+      return this.jdbcTemplate.queryForObject(sql_statement, this.cohortDefinitionMapper);
+  }
+  
+  
+
 }
